+++ contribs/freight/src/main/java/org/matsim/contrib/freight/events/LSPFreightLinkEnterEvent.java
@@ -1,18 +1,23 @@
-package org.matsim.contrib.freight.controler;
+package org.matsim.contrib.freight.events;
 
 import java.util.Map;
 
 import org.matsim.api.core.v01.Id;
 import org.matsim.api.core.v01.events.Event;
+import org.matsim.api.core.v01.events.LinkEnterEvent;
 import org.matsim.api.core.v01.network.Link;
 import org.matsim.api.core.v01.population.Person;
 import org.matsim.contrib.freight.carrier.Carrier;
 import org.matsim.contrib.freight.carrier.CarrierVehicle;
 import org.matsim.vehicles.Vehicle;
 
-public final class LSPFreightLinkLeaveEvent extends Event {
 
-	public static final String EVENT_TYPE = "freight vehicle left link";
+//Achtung: Das darf nicht von LinkEnterEvent erben, weil sonst im CarrierResourceTracker immer wenn ein ein solches event hergestellt wird, der CarrierResourceAgent +
+//denkt ein neues LinkEnterEvent waere geworfen worden -> Endlosschleife
+
+public final class LSPFreightLinkEnterEvent extends Event{
+
+	public static final String EVENT_TYPE = "freight vehicle entered link";
 	public static final String ATTRIBUTE_VEHICLE = "vehicle";
 	public static final String ATTRIBUTE_LINK = "link";
 	public static final String ATTRIBUTE_CARRIER = "carrier";
@@ -24,7 +29,7 @@ public final class LSPFreightLinkLeaveEvent extends Event {
 	private Id<Vehicle> vehicleId; 
 	private Id<Link>linkId; 
 	
-	public LSPFreightLinkLeaveEvent(Id<Carrier>carrierId, Id<Vehicle> vehicleId, Id<Person>driverId, Id<Link>linkId, double time, CarrierVehicle vehicle) {
+	public LSPFreightLinkEnterEvent(Id<Carrier>carrierId, Id<Vehicle> vehicleId, Id<Person>driverId, Id<Link>linkId, double time, CarrierVehicle vehicle) {
 		super(time);
 		this.carrierVehicle = vehicle ;
 		this.carrierId = carrierId;
@@ -67,4 +72,5 @@ public final class LSPFreightLinkLeaveEvent extends Event {
 		attr.put(ATTRIBUTE_DRIVER, this.driverId.toString());
 		return attr;
 	}
+	
 }
