+++ contribs/freight/src/main/java/org/matsim/contrib/freight/events/LSPShipmentDeliveredEventCreator.java
@@ -1,4 +1,4 @@
-package org.matsim.contrib.freight.controler;
+package org.matsim.contrib.freight.events;
 
 import org.matsim.api.core.v01.Id;
 import org.matsim.api.core.v01.events.ActivityEndEvent;
@@ -9,7 +9,6 @@ import org.matsim.contrib.freight.carrier.Carrier;
 import org.matsim.contrib.freight.carrier.FreightConstants;
 import org.matsim.contrib.freight.carrier.ScheduledTour;
 import org.matsim.contrib.freight.carrier.Tour.Delivery;
-import org.matsim.contrib.freight.events.ShipmentDeliveredEvent;
 
 public final class LSPShipmentDeliveredEventCreator implements LSPEventCreator {
