+++ contribs/freight/src/main/java/org/matsim/contrib/freight/events/LSPTourStartEvent.java
@@ -1,4 +1,4 @@
-package org.matsim.contrib.freight.controler;
+package org.matsim.contrib.freight.events;
 
 import java.util.Map;
 
@@ -8,24 +8,23 @@ import org.matsim.api.core.v01.population.Person;
 import org.matsim.contrib.freight.carrier.Carrier;
 import org.matsim.contrib.freight.carrier.CarrierVehicle;
 import org.matsim.contrib.freight.carrier.Tour;
-import org.matsim.contrib.freight.carrier.Tour.End;
+import org.matsim.contrib.freight.carrier.Tour.Start;
 
-public final class LSPTourEndEvent extends Event{
+public final class LSPTourStartEvent extends Event{
 
-	public static final String EVENT_TYPE = "freight tour ended";
+	public static final String EVENT_TYPE = "freight tour started";
 	public static final String ATTRIBUTE_VEHICLE = "vehicle";
 	public static final String ATTRIBUTE_LINK = "link";
 	public static final String ATTRIBUTE_CARRIER = "carrier";
 	public static final String ATTRIBUTE_DRIVER = "driver";
 	public static final String ATTRIBUTE_TOUR = "tour";	
 	
-	
 	private Id<Carrier> carrierId;
 	private Id<Person> driverId;
 	private Tour tour;
 	private CarrierVehicle vehicle;
 	
-	public LSPTourEndEvent(Id<Carrier>  carrierId, Id<Person> driverId, Tour tour, double time, CarrierVehicle vehicle) {
+	public LSPTourStartEvent(Id<Carrier>  carrierId, Id<Person> driverId, Tour tour, double time, CarrierVehicle vehicle) {
 		super(time);
 		this.carrierId = carrierId;
 		this.driverId = driverId;
@@ -49,11 +48,11 @@ public final class LSPTourEndEvent extends Event{
 	public Tour getTour() {
 		return tour;
 	}
-	
+
 	public CarrierVehicle getVehicle() {
 		return vehicle;
 	}
-
+	
 	@Override
 	public Map<String, String> getAttributes() {
 		Map<String, String> attr = super.getAttributes();
